requirements.txt
streamlit==1.31.0
pillow==10.2.0
google-cloud-vision==3.7.0
googletrans==4.0.0rc1
reportlab==4.1.0
PyPDF2==3.0.1
python-dotenv==1.0.0
